+++ src/main/java/com/rcextract/minecord/EventManager.java
@@ -69,9 +69,8 @@ public class EventManager implements Listener {
 		Bukkit.getPluginManager().callEvent(e);
 		if (e.isCancelled()) return;
 		for (User user : users) 
-			if (user.isOnline()) {
+			if (user.isOnline()) 
 				user.getOnlinePlayer().sendMessage(Minecord.applyFormat(user.getName(), user.getNickName(), user.getPlayer().getUniqueId().toString(), e.getMessage(), new Date().toString()));
-			}
 		//Tag detection
 		/*LinkedHashMap<String, Player> segments = new LinkedHashMap<String, Player>();
 		for (String segment : event.getMessage().split("@")) segments.put(segment, null);